Begin of some refactoring
