+++ oggm_edu/__init__.py
@@ -1,7 +1,10 @@
-__version__ = '0.0.1'
+__version__ = "0.1.0b"
 
-import oggm_edu.funcs as legacy_functions
+from oggm_edu.funcs import plot_glacier_graphics, initalize_oggm
 from oggm_edu.glacier import Glacier, SurgingGlacier
 from oggm_edu.glacierBed import GlacierBed
 from oggm_edu.glacierCollection import GlacierCollection
 from oggm_edu.massBalance import MassBalance
+
+# Initialize on import
+initalize_oggm()
